Создал компоненту nearOfferBlock
